+++ src/main/java/wooteco/subway/infrastructure/exception/EmailInvalidBoundaryException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.infrastructure.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
