+++ src/main/java/wooteco/subway/infrastructure/exception/duplicate/LineDuplicatedException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.duplicate;
+package wooteco.subway.infrastructure.exception.duplicate;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
