+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -1,8 +1,8 @@
 package wooteco.subway.member.application;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.duplicate.EmailDuplicatedException;
-import wooteco.subway.exception.not_found.MemberNotFoundException;
+import wooteco.subway.infrastructure.exception.duplicate.EmailDuplicatedException;
+import wooteco.subway.infrastructure.exception.not_found.MemberNotFoundException;
 import wooteco.subway.member.dao.MemberDao;
 import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.member.domain.Member;
