+++ src/main/java/wooteco/subway/path/application/FarePrincipalFinder.java
@@ -27,6 +27,7 @@ public class FarePrincipalFinder {
         if (age < 19) {
             return new TeenagerFarePrincipal();
         }
+
         return new AdultFarePrincipal();
     }
 }
