+++ src/main/java/wooteco/subway/path/domain/fare/FarePrincipal.java
@@ -3,8 +3,6 @@ package wooteco.subway.path.domain.fare;
 import static wooteco.subway.path.application.FarePrincipalFinder.FARE_PER_KM;
 
 public interface FarePrincipal {
-    int DEDUCTED_AMOUNT = 350;
-
     double calculateFare(int distance, long extraFare);
 
     default double calculateOverFare (double baseFare, int distance, int perDistance){
