+++ src/main/java/wooteco/subway/path/domain/fare/TeenagerFarePrincipal.java
@@ -3,7 +3,8 @@ package wooteco.subway.path.domain.fare;
 import static wooteco.subway.path.application.FarePrincipalFinder.*;
 
 public class TeenagerFarePrincipal implements FarePrincipal {
-    private double discountRate = 0.2;
+    private static int DEDUCTED_AMOUNT = 350;
+    private static double DISCOUNT_RATE = 0.2;
 
     @Override
     public double calculateFare(final int distance, final long extraFare) {
@@ -19,6 +20,6 @@ public class TeenagerFarePrincipal implements FarePrincipal {
             fare = calculateOverFare(baseFare, distance - SECOND_OVER_FARE_DISTANCE, 8);
         }
 
-        return fare - ((fare - DEDUCTED_AMOUNT) * discountRate);
+        return fare - ((fare - DEDUCTED_AMOUNT) * DISCOUNT_RATE);
     }
 }
