+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,9 +1,9 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.delete.StationDeleteException;
-import wooteco.subway.exception.duplicate.StationDuplicatedException;
-import wooteco.subway.exception.not_found.StationNotFoundException;
+import wooteco.subway.infrastructure.exception.delete.StationDeleteException;
+import wooteco.subway.infrastructure.exception.duplicate.StationDuplicatedException;
+import wooteco.subway.infrastructure.exception.not_found.StationNotFoundException;
 import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.line.dto.SectionEntity;
 import wooteco.subway.station.dao.StationDao;
